command: Accept command arguments instead of an `exec.Cmd`
In order to spawn a command via our command package the caller has to
pass in an `exec.Cmd`. This calling convention is weird and opens up the
window for unexpected behaviour in case the caller passes in a command
that is initialized with state unexpected by the command package.

Refactor the package to instead accept a set of arguments. Like this,
the assembled `exec.Cmd` is fully controlled by the command package.