command: Add `WithDir()` option
Callers of the command package that wanted to run the command in a
specific directory had to manually set the directory of the `exec.Cmd`
they passed into the command package. This is a weird calling convention
as ideally, the command package should handle the complete setup for the
caller.

Add a new option `WithDir()` to handle this usecase inside of the
command package and convert callers to use it.