datastore: Do not enqueue duplicate update jobs
The reconciler schedules replication jobs for repositories that are not
up-to-date. It creates a new event each time it notices a repo is
outdated. We've seen situations where the job queue gets filled up with
a large amount of replication jobs. This is extremely inefficient and
makes the table grow to unmanageable sizes.

This change will avoid creation of a new update job when there is
already one there that's not being processed yet for the same repo.

Changelog: performance
Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3940