praefect: Add -db-only flag to remove repository
In some situations, most notably Geo replication, a large number of
repos may be present in the DB but not on disk. It's possible to remove
these entries using the `remove-repository` sub-command, but this
creates a risk of accidentally deleting a repo that's present on disk
and triggering permanent dataloss.

This commit adds a new `-db-only` flag to `remove-repository` allowing
admins to be confident that even if an on-disk repo is passed in it
won't be deleted from disk and can be re-tracked easily.

Changelog: added