praefect: Always use replica paths in tests
We currently enabled feature flag `PraefectGeneratedReplicaPaths` half
the time in testing. Several tests for the `remove-repository`
are checking that the repository relative path does not exist, but when
this feature is enabled the replica path will differ, leading us to
check the wrong path.

Update these tests to always use `replicaPath` and confirm that the path
exists prior to removing it as a check to prevent this from recurring.