repository: Fix test writing into Gitaly's repository
One of our tests for `clnoeFromURLCommand()` tries to clone into a
subdirectory of the Gitaly repository itself. This is wrong: all test
data should always end up in a temporary directory.

Fix this by properly cloning into a temporary directory. While at it,
add a call to `cmd.Wait()` to not leak the process.