gittest: Introduce default object hash
The gittest package is currently re-exporting the functionality provided
by the `ObjectHash` structure. Depending on whether we are testing with
SHA1 or SHA256, we export either the implementation details of the
`ObjectHashSHA1` or `ObjectHashSHA256` structure.

Now that we have the hash-specific information neatly encapsulated in
its own structure there is no need to re-export the separate details of
that structure anymore. Instead, refactor the gittest package to only
export a single `DefaultObjectHash` variable that points to one of the
above structures, depending on what we're currently testing.