updateref: Skip another known-flaky test with Git v2.33.0
In 53944694c (updateref: Skip known-flaky test with Git v2.33.0,
2022-07-08) we have started to skip a test that's known to be flaky with
Git v2.33.0 or older. While this test is not flaking anymore, we're now
regularly seeing a similar flake in another test. This new flaky test is
related to cancellation, but it also does the same `Prepare()` step and
expects that to properly lock the references, which is not guaranteed by
Git v2.33.0.

Skip this test as well to decrease flakiness.