git: Hide refs for upload-pack
Currently we are relying on an omnibus setting that sets a git config
value to hide certain hidden refs during transfer. However, we no longer
want to rely on git config files that are external to Gitaly but rather
inject config values we need directly into the command.

We already do this for receive-pack but not for upload-pack. Add the
corresponding config for upload-pack.

Changelog: changed