git: Discern different types of internal references
We've got two different types of internal references: those that are
interesting to the user but shouldn't be written to by them, and then
those that are really only required for our own internal housekeeping
and which should neither be readable nor writeable.

Introduce a new `InternalReferenceType` to mark references accordingly.