smarthttp: Test for exact matches of ref advertisement in InfoRefs tests
The tests for InfoRefs endpoints in the smarthttp test suite verify that
we advertise a subset of expected references. This isn't ideal though
because we cannot check for the absence of specific references like
this.

Refactor the tests to create their test state via an empty repository so
that we end up with a small set of references, only. Like this, we can
migrate tests to instead check the full set of references now that there
are only two or three references anymore instead of hundreds.