smarthttp: Demonstrate that we mistakenly announce hidden refs
The internal references known by Gitaly are not all the same: while some
of them are internal as in read-only, others are internal as in they
should not be visible to the client at all. We don't currently configure
this directly and thus announce them to the client.

Add testcases that demonstrate this misconfiguration.