proto: RPC to check service readiness
The new ReadinessCheck RPC added. It will allow to check the
service readiness and should be triggered before putting traffic
on to the service. In case of the failure the details about an
error and the check name would be returned to the caller.

Part of: https://gitlab.com/gitlab-org/gitlab/-/issues/348174