command: Remove useless check for whether the process is set
After starting an `exec.Cmd` the `Process` field is always set according
to the documentation of it. Given that we only ever access that field
after we have started the command already it's thus not needed to check
whether the field is set or not.

Remove the check to clean up the code.