gittest: Disallow use of seeded repositories with SHA256
It's not currently possible to use our test repositories in combination
with SHA256 given that they're all using SHA1. This makes it easy to
accidentally test with SHA1 even though one thought to be testing with
SHA256.

Detect the use of seeded repositories and fail the test when done with
SHA256 is the default hash.