gittest: Don't set up default parent in WriteCommit
When not explicitly given a set of parents in WriteCommit, then we fall
back to using a well-known commit in our default seed repository. This
may both be unexpected and makes it needlessly complex to use this
function in an empty repository.

Drop the fallback logic and create commits without any parents if not
asked to do so.