git: Expose hash function used by object hashes
Expose the hash function used by our different object hash
implementations so that we can both learn about the hash size and
compute our own object hashes in tests.

Note that this makes it impossible to compare the structure anymore due
to the added function pointers. We thus have to adjust a small set of
sites to not do that anymore.