gittest: Refactor tree-related functionality to be hash-clean
Refactor both RequireTree and the WriteTree tests to both be hash-clean
so that they work with SHA1 and SHA256. Note that we're also dropping
the expected object hash from our tests: we already verify that tree
entries match, so it doesn't feel sensible to check this opaque hash in
the first place.