Change FindTag err when tag not found to NotFound
What
---
Update the error message to be `Not Found` rather then `Internal` when a
tag is not found.

Why
---
We burn through our error budget when a single user sends a bunch of
requests to tags that don't exist because we classify the error as
`Internal`, instead it should be `NotFound`.

Reference: https://gitlab.com/gitlab-org/gitaly/-/issues/4366
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>