operations: Refactor Ruby-compatible argument validation
The validation we have for arguments passed to the UserCreateTag RPC is
still carrying around quite some baggage from the old Ruby days. While
this made sense back in the day to reduce friction when migrating the
implementation of this RPC from Ruby to Go, this does feel somewhat
dated nowadays.

Refactor validation of arguments to return Go-style error messagess and
wrap the error in an `InvalidArgument` error code.