praefect: Add client into coordinator
The coordinator will make use of the gitlab client to call the features
endpoint to get the enabled and disabled feature flags. Prepare for this
by providing the client in the struct. Use the FeatureGetter interface
so we can pass in an empty FeatureGetter in case we don't have the
configuration necessary to create a gitlab client.