Change FindTag err when tag not found to NotFound
What
---
- Update the error message to be `Not Found` rather then `Internal` when
  a tag is not found behind a feature flag.

Why
---
We burn through our error budget when a single user sends a bunch of
requests to tags that don't exist because we classify the error as
`Internal`, instead it should be `NotFound`.

We need to set this behind a feature flag to prevent any deployment
issues because the [rails application](https://bit.ly/3onH0x6) expects
`Internal` err. In
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/93009
we are adding support for `NotFound`.

Changelog: changed
Reference: https://gitlab.com/gitlab-org/gitaly/-/issues/4366
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>