proto: Remove unused `Repack` field from FetchIntoObjectPoolRequest
The `FetchIntoObjectPoolRequest` messages has a `Repack` field that
supposedly controls whether the repository should or should not be
repacked. This flag is completely ignored though and has always been
ignored. It also isn't ever set by the Rails client.

Remove the field without replacement for the time being. It seems like
there is no usecase for it right now and it's bad to keep a no-op flag
around in our code. Furthermore, we will want to fully move repository
housekeeping into Gitaly eventually anyway.