gitlab: Expose FeatureGetter interface
Add a FeatureGetter interface that only includes the Features method. In
the case that we can't create a gitlab client, we don't want to fail
startup but just use some kind of noop client. There is the MockClient,
but we wouldn't want to use a MockClient in application code. Nor would
it be ideal to create a new EmptyClient since the API surface would be
unnecessarily large since we would only be calling the Features method.

With a separate FeatureGetter, we can add an empty feature getter that
returns nothing instead of having to use the MockClient