gittest: Iterate on delta islands test
In 04b39de9f (objectpool: Fix test for usage of delta islands,
2022-07-13) we have refactored the tests for delta islands so that they
work as expected in the context of object pools. The result still isn't
quite satisfactory though as it puts too many assumptions about how the
repacking function is supposed to work in the test itself. Thus, the
test is about to break as soon as we start pruning pool references.

Refactor the test again. This time, we acknowledge that the repo we're
doing the setup in can be different from the one we're ultimately
repacking to account for the behaviour of FetchIntoObjectPool.