gittest: Allow writing arbitrary references in WriteCommit
The WriteCommit helper function has an option to write the newly written
commit to a specific branch. It's not possible though to write to an
arbitrary reference that is not a branch.

Add another option `WithReference()` that allows callers to chose custom
reference prefixes so that they can write non-branch-references.