ruby: Fix dependencies in Gemfile.lock
With 4da75e5814680fe0d657bb734099527c74b76905 we attempted to update
actionview to resolve a security issue. However, in doing so we updated
the version of actionpack and its dependencies, but not actionview
itself. This has led to builds emitting the following warning:

  Your lockfile doesn't include a valid resolution.
  You can fix this by regenerating your lockfile or trying to manually editing the bad locked gems to a version that satisfies all dependencies.
  The unmet dependencies are:
  * actionview (= 6.1.5.1), depended upon actionpack-6.1.5.1, unsatisfied by actionview-6.1.4.7
  * activesupport (= 6.1.5.1), depended upon actionpack-6.1.5.1, unsatisfied by activesupport-6.1.4.7

Update Gemfile.lock to update all related dependencies to matched
versions and stop this error.