proxy: Get rid of shared state via test suite
The extended proxy handler tests use a testing suite to group together a
set of tests that have similar setup. While this helps reduce the time
used to execute those tests, the shared state really makes it hard to
understand what exactly is going on in those tests. Furthermore, it
makes it harder than necessary to iterate on the implementation of a
specific test without also impacting the other tests.

Split out the test suite's setup into a `setupProxy()` function and
convert the tests into standalone tests that don't use a shared test
suite anymore.