objectpool: Remove feature flag guarding heuristical repo maintenance
In 410295810 (objectpool: Switch to use OptimizeRepository for pool
repos, 2022-07-12) we have migrated maintenance of object pools in
`FetchIntoObjectPool()` to use the same repository maintenance as we use
for normal repositories. This deduplicates the logic, makes sure we keep
more data structures up to date in object pools and also unifies nightly
and normal maintenance for pool repositories.

We have rolled out this feautre flag to production without any issues
observed so far. So let's remove the feature flag and unconditionally
enable heuristical repository maintenance for object pools.

Changelog: changed