Remove legacy go.mod and go.sum files from _build directory
The build process left a go.mod and go.sum file in the _build directory
in the past. This causes problems these days when attempting to embed
the auxiliary binaries into Gitaly binary as they are considered to be
in a different module due to this. Remove the legacy files prior to
building Gitaly.