config: Generisize functions that create and prune temp directories
Pruning directories is a common operation. Currently it's used for
pruning Gitaly runtime directories. However, Cgroup directories are made
according to the same convention as Gitaly runtime directories in the
form of gitaly-<pid>. Instead of reproducing the logic, simply rename
the function so it can be reused.

To do this, we also need to encapsulate the logic that constructs the
name of a process directory and the logic that parses out the pid from
such a directory into functions.