ssh: Simplify handling of `GlRepository()` for logging purposes
When receiving an SSHUploadPack request we log some information included
in the request. Part of that information is the information about the
GitLab repository the request has been created for, which we laboriously
extract from the request while paying attention to not dereference any
`nil` pointers.

Simplify the code to insetad use the Protobuf's getter functions, which
already know to handle `nil` pointers for us.