objectpool: Update test to exercise transactional voting
One of our tests for FetchIntoObjectPool exercises the use of hooks in
that RPC. This test is a historic leftover from the past where we didn't
do transactional voting at all in that function, which shows via one of
the comments we still have that requires that we don't ever execute the
reference-transaction hook.

This is no longer true though, and thus the test doesn't make sense in
its current form. Refactor it to instead assert that we're voting with
the expected inputs.