ssh: Verify InvalidArgument errors in ReceivePack test
While we verify the error code returned by SSHReceivePack in our tests
for argument validation, we don't verify that validation indeed failed
due to the reason we expect it to.

Refactor the tests to also verify the error message of returned errors.
This surfaces that in one of our tests for invalid arguments we failed
to build `gitaly-ssh` and thus tests have failed because of the missing
executable.

While at it, stop using a seeded repository: we're not hitting the repo
anyway.