ssh: Recognize cancelled pushes in SSHReceivePack
Recently we've fixed cases where fetches which are cancelled by the user
will be logged as `Internal` errors  and thus easily impact our SLOs.
This shortcoming has been fixed has been fixed via a2fee3dd5 (ssh:
Improve errors for fetches canceled by the user, 2022-07-20), but now
that we started to report errors in SSHReceivePack we are likely to hit
the same issue there.

Let's apply the same fix to SSHReceivePack and special-case this
specific error.