praefect: Add 'track-repositories' subcommand
A somewhat common way to migrate to Gitaly Cluster it to copy the
contents of an existing storage onto one node of the cluster and
manually track each repository. This is convenient for instances where
the migration must happen as a single 'big bang' with some downtime, vs
the zero-downtime, but much slower method of migrating via the API.

Currently admins can script the `track-repository` to run in a loop when
ingesting large numbers of repositories, but this is inconvenient and
error-prone. To better support this use case, a way to import
repositories in-bulk into Praefect is needed.

This commit add a new `track-repositories` subcommand that takes a JSON
file as input and adds each entry in a single execution.

Changelog: added