praefect: Make track-repository output consistent
Currently most output from the `track-repository` subcommand is made via
the commands writer, but a subset is done via logger. This leads to an
inconsistent format where some output is plaintext, while other messages
are in JSON, which is difficult to read.

Let's make this consistent by making all non-debug output use the
writer, which will be simpler to parse for a one-off CLI command.