ref: Split out code related to the FindTag RPC
The FindTag RPC is implemented alongside a set of other RPCs in the
`refs.go` file. It's thus a tad hard to find and doesn't conform to our
usual best practice, where each RPC is implemented in its own code unit.

Move the code into a separate file to make it easier to locate.