git: Remove feature flag to hide refs in git-upload-pack(1)
With commit 4a789524c (git: Don't advertise internal references via
git-upload-pack(1), 2022-07-19) we have started to hide a subset of
internal references in git-upload-pack(1) so that these aren't
advertised at all anymore. When this change was rolled out to canary
initially we observed a short CPU burst and increase in writes on that
Gitaly node. This was totally unexpected given that the change was
supposed to only mirror what we have already configured in Omnibus
anyway, so it shouldn't have caused any visible change in behaviour. We
thus decided to retroactively introduce a feature flag for the change
via bc34b47ac (git: Introduce feature flag for hiding refs in
git-upload-pack(1), 2022-07-26).

As it turns out though, the observed behaviour was completely unrelated
to the change at hand, and we have subsequently rolled out the flag to
production without any observed weirdnesses as initially expected. So
let's remove the feature flag again.