repository: Resolve revisions passed to the WriteRef RPC
The WriteRef RPC gets as input a reference name that is to be updated as
well as the old and new revisions that the reference should be updated
from respectively to. It's not entirely clear what these revisions
contain: while one would typically expect it to only ever contain object
IDs, that may or may not be the case.

Resolve the revision to an object ID before invoking git-update-ref(1)
with them. This will ease the transition when we migrate the updateref
package to only ever accept object IDs as input, but it will also give
us better indicators when the revision does not exist in the repository.