objectpool: Verify object IDs when rescuing dangling objects
When we rescue dangling objects in an object pool we execute git-fsck(1)
to find out about any objects which are not currently referenced. We
don't perform any sanity checks though on the output of git-fsck(1): the
only thing we do is to split the string, and then pass whatever data we
have obtained to git-update-ref(1).

Verify that the obtained data is actually a conforming object hash by
parsing the object ID. This also prepares for an upcoming change where
the updateref package will start to accept `git.ObjectID`s instead of
bare strings.