gittest: Create test data at runtime in the repository suite
The shared repository suite that is used to verify both the localrepo
and remoterepo implementations currently uses a toggle to tell the
implementation-specific setup function whether the repository should be
seeded or not. This doesn't make a whole lot of sense: setup of the test
data really should be handled by the suite itself.

Refactor the code and create the test data at runtime. This also enables
us to easily test against SHA256.