gittest: Disable use of CreateRepository with SHA256 in general
We have already disabled the use of CreateRepository with seeded
repositories in 9ef8ff076 (gittest: Disallow use of seeded repositories
with SHA256, 2022-07-21) when SHA256 is in use. This isn't broad enough
though given that this function uses the `CreateRepository()` RPC behind
the scenes which doesn't yet support SHA256 at all, so it doesn't even
work to use this function with unseeded repositories.

Let's add another test failure that is specific to the unseeded-repo
case to avoid accidental use of this function in SHA256 tests.